Added a pop up alert when deleting task

+++ src/component/SingleTodoList.js
@@ -12,11 +12,14 @@ const SingleTodoList = ({title, isDone,onEdit,id,history}) =>{
 			{isDone ? "This has been completed" : "Has yet to be finished"}		
 			</div>
 			<div className="d-flex justify-content-between align-items-end mt-5">
-			<button className="bt btn-link text-black " onClick={()=>{
+			<button className="bt btn-link text-black " onClick={()=>{window.confirm("Are you sure?") &&
 				axios.delete(`/api/list/${id}`).then(()=>{
 				history.push('/')
-				})
+
+			})
+
 			}}>Delete</button>
+
 			<button className="bt btn-link text-black" onClick={onEdit}>Edit</button>
 			</div>
 		</div>
@@ -26,3 +29,6 @@ const SingleTodoList = ({title, isDone,onEdit,id,history}) =>{
 }
   
 export default withRouter(SingleTodoList)
+
+
+//alert("Are you sure to delete this task?")
